requirements.txt
jsonschema==3.2.0
PyYAML==6.0
requests==2.28.1